refactor: extract data models into models.py
